Account for abi-dump failure.

+++ ci/kokoro/docker/check-abi.sh
@@ -53,12 +53,15 @@ if [[ "${UPDATE_ABI}" == "yes" ]]; then
     -lver "expected" \
     -o "${ACTUAL_DUMP_PATH}" &&
     gzip -c "${ACTUAL_DUMP_PATH}" >"${EXPECTED_DUMP_PATH}.tmp"
-  if zdiff "${EXPECTED_DUMP_PATH}.tmp" "${EXPECTED_DUMP_PATH}"; then
-    rm -f "${EXPECTED_DUMP_PATH}.tmp"
-  else
-    mv -f "${EXPECTED_DUMP_PATH}.tmp" "${EXPECTED_DUMP_PATH}"
+  exit_status=$?
+  if [[ $exit_status -eq 0 ]]; then
+    if zdiff "${EXPECTED_DUMP_PATH}.tmp" "${EXPECTED_DUMP_PATH}"; then
+      rm -f "${EXPECTED_DUMP_PATH}.tmp"
+    else
+      mv -f "${EXPECTED_DUMP_PATH}.tmp" "${EXPECTED_DUMP_PATH}"
+    fi
   fi
-  exit $?
+  exit $exit_status
 fi
 
 io::log "Checking ABI of ${LIBRARY}"
